# cmd/api/docs/swagger.yaml
definitions:
  domain.Address:
    properties:
      address_id:
        type: integer
      area:
        type: string
      city:
        type: string
      fname:
        type: string
      house:
        type: string
      landmark:
        type: string
      lname:
        type: string
      phone_number:
        type: integer
      pincode:
        type: integer
      user_id:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server  server. You can visit the GitHub repository
    at https://github.com/prasanth-pn/clean-code-architecture-ecommerce
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin ecommerce API
  version: "1.0"
paths:
...
  /user/add/address:
    post:
      operationId: AddAddress for user
      parameters:
      - description: Address
        in: body
        name: Address
        required: true
        schema:
          $ref: '#/definitions/domain.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: AddAddress for user
      tags:
      - ADDRESSMANAGEMENT
...
  /user/edit/address:
    get:
      operationId: GetAddress for user
      parameters:
      - description: enter the address_id
        in: query
        name: address_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: GetAddress for user
      tags:
      - ADDRESSMANAGEMENT
...
  /user/list/address:
    get:
      operationId: ListAddress for user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: AddAddress for user
      tags:
      - ADDRESSMANAGEMENT
...
  /user/update/address:
    patch:
      operationId: UpdateAddress for user
      parameters:
      - description: Address
        in: body
        name: Address
        required: true
        schema:
          $ref: '#/definitions/domain.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: UpdateAddress for user
      tags:
      - ADDRESSMANAGEMENT
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"